Fix some Issues and refactored the code

+++ password_manager.py
@@ -1,10 +1,8 @@
-from cryptography.hazmat.primitives import hashes
-from cryptography.hazmat.primitives.kdf.pbkdf2 import PBKDF2HMAC
-from cryptography.hazmat.primitives import serialization
-from cryptography.fernet import Fernet
-import base64
 import os
 import logging as lg
+import pyperclip
+import functions
+from prettytable import PrettyTable
 
 lg.basicConfig(level=lg.DEBUG,
                     format="%(asctime)s - %(levelname)s - %(message)s",
@@ -18,16 +16,6 @@ class PasswordManager:
         self.master_password = master_password.encode()
         self.key = key
 
-    def encrypt_password(self, password):
-        fernet = Fernet(self.key)
-        encrypted_password = fernet.encrypt(password.encode())
-        return encrypted_password
-
-    def decrypt_password(self, encrypted_password):
-        fernet = Fernet(self.key)
-        decrypted_password = fernet.decrypt(encrypted_password).decode()
-        return decrypted_password
-
     def clear_screen(self):
         os.system("cls" if os.name == "nt" else "clear")
 
@@ -40,52 +28,67 @@ class PasswordManager:
         print("3. Delete password")
         print("4. Create new password")
         print("5. Update password")
-        print("6. Exit")
+        print("6. Delete all passwords")
+        print("7. Exit")
         print("----------------")
         print("Enter your choice: ", end="")
 
-    # TODO: Check if password meets basic criteria
-    def check_password(self, password):
-        # length between 6 and 64
-        # if its utf-8?
-        # its a letter or a code?
-        pass
-
     def create_password(self, website, username, password):
-        encrypted_password = self.encrypt_password(password)
-        self.db.insert_password(website, username, encrypted_password)
+        stored_item = self.db.view_password(website, username)
+        
+        if stored_item == -1:
+            encrypted_password = functions.encrypt_password(self.key, password)
+            self.db.insert_password(website, username, encrypted_password)
+        else:
+            print("Password already saved, if you want to update it use that option.")
 
-    # TODO: copy to the clipboard directly
     def view_password(self, website, username):
-        stored_password = self.db.view_password(website, username)[3]
-        print(self.decrypt_password(stored_password))
-        lg.debug(f"Password for {website} and {username} is {self.decrypt_password(stored_password)}")
+        stored_password = self.db.view_password(website, username)
+        if stored_password != -1:
+            pyperclip.copy(functions.decrypt_password(self.key, stored_password)[3])
+            print("Password copied to the clipboard")
+            lg.debug(f"Password for {website} and {username} is {functions.decrypt_password(self.key, stored_password)[3]}")
+        else:
+            print("Password not found")
+            lg.debug(f"Password for {website} and {username} not found")
 
-    def delete_password(self, website, username): # add password?
-        self.db.delete_password(website, username)
+    def delete_password(self, website, username):
+        if self.db.view_password(website, username) != -1:
+            self.db.delete_password(website, username)
+            print("Password deleted succesfully")
+        else:
+            print("Password not found")
 
     def show_all_passwords(self):
         all_passswords = self.db.show_all_passwords()[1:]
+
         if len(all_passswords) == 0:
-            print("No passwords stored")
-        else:
-            for _, website, username, password in all_passswords:
-                print(website, username, password, sep=" || ")
+            print("No password saved")
+            return
+
+        t = PrettyTable(['Website', 'Username'])
+        for _, website, username, _ in all_passswords:
+            t.add_row([website, username])
+
+        print(t)
 
     def update_password(self, website, username, password, newpassword):
-        encrypted_password = self.encrypt_password(password)
-        encrypted_newpassword = self.encrypt_password(newpassword)
+        encrypted_newpassword = functions.encrypt_password(self.key, newpassword)
 
         # Check if password exists and its valid
-        stored_password = self.db.view_password(website, username)[3]
+        stored_password = self.db.view_password(website, username)
 
-        if self.decrypt_password(stored_password) == password:
-            self.db.modify_password(website, username, encrypted_newpassword)
-            print('Password succesfully modified')
-            lg.debug(f"Password for {website} and {username} was updated to {self.decrypt_password(encrypted_newpassword)}")
+        if stored_password != -1:
+            if functions.decrypt_password(self.key, stored_password[3]) == password:
+                self.db.modify_password(website, username, encrypted_newpassword)
+                print('Password succesfully modified')
+                lg.debug(f"Password for {website} and {username} was updated to {functions.decrypt_password(self.key, encrypted_newpassword[3])}")
+            else:
+                print('Passwords does not match')
+                lg.debug(f"Password for {website} and {username} was not updated")
         else:
-            print('Passwords does not match')
-            lg.debug(f"Password for {website} and {username} was not updated")
+            print('Account not found')
+            lg.debug(f"Password for {website} and {username} was not found")
 
     def action(self, choice):
         if choice == 1:
@@ -101,29 +104,37 @@ class PasswordManager:
         elif choice == 4:
             website = input("Enter website: ")
             username = input("Enter username: ")
-            password = input("Enter password: ")
+            password = functions.read_password()
             self.create_password(website, username, password)
         elif choice == 5:
             website = input("Enter website: ")
             username = input("Enter username: ")
-            password = input("Enter current password: ")
-            newpassword = input("Enter new password: ")
+            password = functions.read_password(type="current")
+            newpassword = functions.read_password(type="new")
             self.update_password(website, username, password, newpassword)
-        else:
-            print("Invalid choice")
+        elif choice == 6:
+            self.db.delete_all()
+
+    def valid_menu_input(self, menu_input):
+        return menu_input.isdigit() and int(menu_input) <= 7 and int(menu_input) >= 1
 
     def menu(self):
         self.clear_screen()
         while True:
             self.print_menu()
-            choice = int(input())
+            choice = input()
 
-            if choice == 6:
+            if not self.valid_menu_input(choice):
+                print("Invalid choice, try again.")
+                continue
+
+            choice = int(choice)
+
+            if choice == 7:
                 self.db.close_db()
                 break
 
             try:
                 self.action(choice)
             except Exception as e:
-                print(f"Error in {choice}: {e}")
                 lg.error(f"Error in {choice}: {e}")
